+++ airsync-client/build.gradle
@@ -1,14 +1,12 @@
-apply plugin: 'java'
-apply plugin: 'kotlin'
 
 jar {
     from {
         configurations.runtime.collect { it.isDirectory() ? it : zipTree(it) }
         configurations.compile.collect { it.isDirectory() ? it : zipTree(it) }
     }
-    destinationDir = file("$rootDir/build/client/bin")
+    destinationDir = file("$buildDir/bin")
 
-    manifest { attributes 'Main-Class': 'ffc.airsync.client.Main' }
+    manifest { attributes "Main-Class": "ffc.airsync.client.MainKt" }
 
     exclude 'META-INF/*.RSA'
     exclude 'META-INF/*.SF'
@@ -17,53 +15,24 @@ jar {
     exclude 'META-INF/NOTICE*'
     exclude 'META-INF/LICENSE*'
     exclude 'about.html'
-
-}
-
-sourceCompatibility = 1.8
-
-compileKotlin {
-    kotlinOptions.jvmTarget = "1.8"
-}
-compileTestKotlin {
-    kotlinOptions.jvmTarget = "1.8"
 }
 
 dependencies {
-    def jettyVersion = "9.4.1.v20170120"
-    def jerseyVersion = "2.25.1"
     def jdbiVersion = "3.1.0"
     def retrofitVersion = "2.3.0"
 
-    implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation project(':airsync-mysql-log-reader')
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-    implementation "org.eclipse.jetty:jetty-server:$jettyVersion"
-    implementation "org.eclipse.jetty:jetty-servlet:$jettyVersion"
-    implementation "org.eclipse.jetty:jetty-http:$jettyVersion"
-    implementation "org.glassfish.jersey.containers:jersey-container-servlet-core:$jerseyVersion"
-    implementation "org.glassfish.jersey.containers:jersey-container-jetty-servlet:$jerseyVersion"
-    implementation "org.glassfish.jersey.media:jersey-media-json-jackson:$jerseyVersion"
-    implementation "org.elasticsearch.client:transport:6.2.1"
-    implementation "org.eclipse.jetty:jetty-http:$jettyVersion"
-    implementation "org.eclipse.jetty.websocket:websocket-client:9.4.8.v20171121"
-    implementation "org.jdbi:jdbi3-core:$jdbiVersion"
-    implementation "org.jdbi:jdbi3-kotlin:$jdbiVersion"
-    implementation "org.jdbi:jdbi3-kotlin-sqlobject:$jdbiVersion"
-    implementation "com.squareup.retrofit2:retrofit:$retrofitVersion"
-    implementation "com.squareup.retrofit2:converter-gson:$retrofitVersion"
-    implementation "com.google.code.gson:gson:2.8.1"
-    implementation "com.google.firebase:firebase-admin:5.11.0"
-    implementation "com.google.firebase:firebase-core:15.0.0"
-    implementation "com.google.firebase:firebase-messaging:15.0.0"
-    implementation "com.google.gdata:core:1.47.1"
-    implementation "com.fatboyindustrial.gson-jodatime-serialisers:gson-jodatime-serialisers:1.6.0"
-    implementation "com.github.piruin:retrofit-dsl:1.0"
-    implementation "args4j:args4j:2.33"
-    implementation "mysql:mysql-connector-java:5.1.38"
-    implementation "me.piruin:geok-gson:1.0.0-alpha"
-    implementation "commons-codec:commons-codec:1.9"
-    implementation "com.github.ffc-nectec:entities:1.0.0-alpha2"
+    compile fileTree(dir: 'libs', include: ['*.jar'])
+    compile project(':provider')
 
-    testImplementation "junit:junit:4.12"
+    compile "org.elasticsearch.client:transport:6.2.1"
+    compile "org.jdbi:jdbi3-core:$jdbiVersion"
+    compile "org.jdbi:jdbi3-kotlin:$jdbiVersion"
+    compile "org.jdbi:jdbi3-kotlin-sqlobject:$jdbiVersion"
+    compile "com.squareup.retrofit2:retrofit:$retrofitVersion"
+    compile "com.squareup.retrofit2:converter-gson:$retrofitVersion"
+
+    compile "com.github.piruin:retrofit-dsl:1.0"
+    compile "args4j:args4j:2.33"
+    compile "mysql:mysql-connector-java:5.1.38"
+    compile "com.github.ffc-nectec:entities:$entitiesVersion"
 }
