--- airsync-client/src/main/kotlin/ffc/airsync/client/webservice/ApplicationConfig.java
@@ -1,35 +1,0 @@
-/*
- * Copyright (c) 2561 NECTEC
- *   National Electronics and Computer Technology Center, Thailand
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package ffc.airsync.client.webservice;
-
-import org.glassfish.jersey.server.ResourceConfig;
-import org.glassfish.jersey.server.filter.RolesAllowedDynamicFeature;
-
-
-public class ApplicationConfig extends ResourceConfig {
-
-    public ApplicationConfig() {
-        packages("ffc.airsync.client.webservice");
-
-        //store(ApiWebSocks.class);
-        //store(JacksonFeature.class);
-        //register(RolesAllowedDynamicFeature.class);
-        //store(CsrfProtectionFilter.class);
-    }
-}
-
