+++ build.gradle
@@ -11,22 +11,40 @@ buildscript {
     }
 }
 
-
+ext {
+   entitiesVersion = '1.0.0-alpha2'
+}
 
 
 allprojects {
-    task wrapper(type: Wrapper) {
-        gradleVersion = '4.8'
-        distributionUrl = distributionUrl.replace("bin", "all")
-    }
+    apply plugin: 'idea'
 
     repositories {
         mavenCentral()
+        google()
         jcenter()
         maven { url 'https://jitpack.io' }
         maven { url "https://dl.bintray.com/blazei/maven" }
-        maven { url "https://maven.google.com" }
     }
+}
+
+subprojects {
+    apply plugin: 'kotlin'
+
+    sourceCompatibility = 1.8
+    compileKotlin { kotlinOptions.jvmTarget = "1.8" }
+    compileTestKotlin { kotlinOptions.jvmTarget = "1.8" }
+
+    dependencies {
+        implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+
+        testImplementation "junit:junit:4.12"
+    }
+}
+
+task wrapper(type: Wrapper) {
+    gradleVersion = '4.8'
+    distributionUrl = distributionUrl.replace("bin", "all")
 }
