+++ implement/airsync-local-web/build.gradle
@@ -1,0 +1,17 @@
+dependencies {
+    def jettyVersion = "9.4.1.v20170120"
+    def jerseyVersion = "2.25.1"
+
+    implementation project(":interface:airsync-ui")
+    implementation project(":interface:airsync-notification")
+    implementation "com.github.ffc-nectec:entities:$entitiesVersion"
+
+    implementation "org.eclipse.jetty:jetty-server:$jettyVersion"
+    implementation "org.eclipse.jetty:jetty-servlet:$jettyVersion"
+    implementation "org.eclipse.jetty:jetty-http:$jettyVersion"
+    implementation "org.glassfish.jersey.containers:jersey-container-servlet-core:$jerseyVersion"
+    implementation "org.glassfish.jersey.containers:jersey-container-jetty-servlet:$jerseyVersion"
+    implementation "org.glassfish.jersey.media:jersey-media-json-jackson:$jerseyVersion"
+
+    testImplementation "junit:junit:4.12"
+}
