+++ implement/airsync-local-web/src/main/java/ffc/airsync/localweb/debug.kt
@@ -1,0 +1,8 @@
+package ffc.airsync.localweb
+
+val debug = System.getenv("FFC_DEBUG")
+
+inline fun printDebug(infoDebug: String) {
+    if (debug == null)
+        println(infoDebug)
+}
