+++ implement/airsync-mysql-log-reader/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+    implementation 'net.sourceforge.javacsv:javacsv:2.0'
+    implementation 'commons-codec:commons-codec:1.11'
+}
