+++ implement/airsync-mysql-log-reader/src/main/java/th/in/ffc/airsync/logreader/filter/QueryFilter.java
