+++ implement/airsync-mysql-log-reader/src/test/resources/ReadTextFileRT.txt
