+++ interface/airsync-notification/src/main/java/ffc/airsync/notification/Notification.kt
@@ -1,0 +1,8 @@
+package ffc.airsync.notification
+
+interface Notification {
+
+    fun onTokenChange(callback: (String) -> Unit)
+
+    fun onReceiveDataUpdate(callback: (type: String, id: String) -> Unit)
+}
