+++ interface/airsync-ui/src/main/java/ffc/airsync/ui/AirSyncUi.kt
@@ -1,0 +1,6 @@
+package ffc.airsync.ui
+
+interface AirSyncUi {
+
+    fun start()
+}
